// code/frontend/frp-fe/src/api/models/AccConnectorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AccCredentialFieldDto } from './AccCredentialFieldDto';
export type AccConnectorDto = {
    type?: string;
    credentialFields?: Array<AccCredentialFieldDto>;
};